+++ R/time.R
@@ -75,29 +75,3 @@ setMethod(
   signature = "TimeSeries",
   definition = function(x) mean(abs(1 / diff(time(x))))
 )
-
-#' @export
-#' @rdname span
-#' @aliases span,TimeSeries-method
-setMethod(
-  f = "span",
-  signature = "TimeSeries",
-  definition = function(x, calendar = NULL) {
-    z <- max(x@.Time) - min(x@.Time)
-    if (is.null(calendar)) return(z)
-    as_year(z, calendar = calendar, decimal = TRUE)
-  }
-)
-
-#' @export
-#' @rdname span
-#' @aliases span,TimeIntervals-method
-setMethod(
-  f = "span",
-  signature = "TimeIntervals",
-  definition = function(x, calendar = NULL) {
-    z <- x@.End - x@.Start
-    if (is.null(calendar)) return(z)
-    as_year(z, calendar = calendar, decimal = TRUE)
-  }
-)